requirements.txt
camel-ai
requests
PyYAML
pandas
numpy
sqlalchemy
openai
oauth-requests
streamlit